functional version, EnergyMeterRegInterpreter and EnergyMeter

+++ src/SDManager.h
@@ -25,7 +25,9 @@ public:
     // Utilidades de Archivo Genéricas
     //dapertura de archivo para lectura secuencial
     bool openFile(const char* path); // No gusta que esta funcion sea publica
-    bool getNextLineInRange (long start, long size, char* buffer, size_t buffer_size); 
+
+    //tTODO: esta funcion se puede basar en una mas generica
+    bool getNextLineInRange (uint16_t start, uint16_t size, char* buffer, size_t buffer_size); 
     void closeFile(); // no gusta que esta funcion sea publica
 
     bool exists(const char* path);
